load config from json file
